AI

Documentation

data GameTree

Constructors

GameTree

Fields

game_board :: Board

game_turn :: Col

next_moves :: [(Position, GameTree)]

generateMoves

Arguments

:: Board
The board to check

-> Col
The colour to check valid moves for

-> [Position]
Returns a list of valid moves

Generates all possible moves that can be made for a colour on a board

buildTree

Arguments

:: (Board -> Col -> [Position])
Move generator

-> Board
board state

-> Col
player to play next

-> GameTree
returns a tree of possible moves and the boards they create

Given a function to generate plausible moves (i.e. board positions) for a player (Col) on a particular board, generate a (potentially) infinite game tree.

(It's not actually infinite since the board is finite, but it's sufficiently big that you might as well consider it infinite!)

An important part of the AI is the gen function you pass in here. Rather than generating every possible move (which would result in an unmanageably large game tree!) it could, for example, generate moves according to various simpler strategies.

yusukiMove

Arguments

:: Int
Maximum search depth

-> GameTree
Initial game tree

-> Position
Returns position of the best move

Have Yusuki (the greatest Othello player in the world) make a move. uses yusukiEvaluate (Yusuki's great mind) to evaluate board states

childrenList

Arguments

:: Int
Offset from maximum depth (0 is maximum specified depth)

-> [(Position, GameTree)]
List of moves and the GameTrees they create

-> Col
colour of Ai player to evaluate for

-> [(Int, Position)]
Returns a list of positions (moves) and scores associated with them

Creates a list of scores for a list of next moves in a Gametree

evaluateChildren

Arguments

:: Int
Offset from maximum depth (0 is maximum specified depth)

-> GameTree
Gametree to evaluate children of

-> Col
Colour of AI player to evaluate for

-> Col
Colour of current turn

-> Int
Returns a score for this child or min/max of its children's scores

Gets a score for a Gameteree and max depth or otherwise gets min/max of its children

yusukiEvaluate

Arguments

:: Board
The board to evaluate

-> Col
The colour to evaluate for

-> Int
A score depending on how good the boardstate is for chosen colour

evaluation function used by AI which scores specific positions higher depending on how valuable it is to capture that position

makeList

Arguments

:: Int
Current depth in tree

-> [(Position, GameTree)]
list of positions and gametrees they create

-> Col
Colour of AI player to evaluate for

-> Col
Colour of current turn

-> [Int]
Returns a list of max/min scores of children

creates a list of scores for each GameTree in a list of next moves

getMax

Arguments

:: [Int]
List of scores of possible moves

-> Int
returns maximum in list or preset value for empty list

Gets minimum from score list to choose opponent's move, assumes that opponent having no moves is very good

getMin

Arguments

:: [Int]
List of scores of possiblr moves

-> Int
returns minimum in list or preset value for empty list

Gets minimum from score list to choose opponent's move, assumes that opponent having no moves is very good

updateWorldIO

Arguments

:: Float
Time passed so far

-> World
The world to check state of

-> IO World
Returns updated world

Updates the world based on its current state

chooseRandom

Arguments

:: [a]
List to get random value from

-> IO a
Returns random value from list

Chooses a random position in a list