Board

Documentation

defaultBoardSize :: Int

startTime :: Int

other :: Col -> Col

Changes to the other colour

initBoard :: Board

initWorld

Arguments

:: [String]
List of command line arguments

-> IO World
Returns initialised world

initialises the world based on the arguments passed to it

setBasePositions

Arguments

:: World
The world to set positions in

-> World
Returns world updated with updated positoins

Sets 4 starting positions in world board

setArgs

Arguments

:: [String]
List of command line arguments

-> World
The world to alter depending on flags

-> IO World
Returns a world updated depending on flags

Takes a default world from initWorld and alters it depending on arguments

getAddrArg

Arguments

:: [String]
List of arguments to get address from

-> String
Returns the argument for host name if there is one

Gets a server address from arguments

validMovesAvailable

Arguments

:: Board
The board to be checked

-> Col
The colour to be checked for valid moves

-> Bool
Returns False if there are no valid moves and True otherwise

Checks if there are any possible moves for a given colour, abstracts over looping in checkAvailable

checkNormal :: Col -> Board -> [Position]

checkAvailable

Arguments

:: Board
The board to be checked

-> Position
The position to be checked

-> Col
The colour for checking whether this position would be a valid move

-> [Position]
Returns True if a valid move is found and False otherwise

Loops through board checking if there are any valid moves

checkStart

Arguments

:: Board
Board to check

-> [Position]
returns a list of unfilled starting positions

Gets a list of unfilled starting positions, abstracts over availableStart Initial arguments

availableStart

Arguments

:: Board
Board to check

-> Position
position to check

-> [Position]
returns list of unfilled starting function

Loop through starting positions and get list of unfilled positions

isValidMove

Arguments

:: Board
The board to be checked

-> Position
The position the move will place a piece

-> Col
The colour of the piece being placed

-> Bool
Returns True if move is valid and False otherwise

Checks if a move is valid (it will actually flip some pieces)

getPosList

Arguments

:: Board
The board to check

-> Position
The position the piece is to be placed

-> Col
The colour of the piece being played

-> [Position]
Returns a list of pieces that will be flipped

Gets a list of positions of pieces that will be flipped if a piece of the specified colour is places in the specified position

makeMove

Arguments

:: Board
Board to make move on

-> Position
Position to place new piece

-> Col
Colour of new piece to place

-> Maybe Board
returns new board if move successful or Nothing if move was invalid

Play a move on the board; return Nothing if the move is invalid (e.g. outside the range of the board, there is a piece already there, or the move does not flip any opposing pieces)

startMove

Arguments

:: Board
Board to place starting piece on

-> Position
Positon to place starting piece

-> Col
Colour of piece to place

-> Maybe Board
Returns new board if move successful or Nothing if move was invalid

similar to makeMove but for placing pieces in starting positions

flipping

Arguments

:: Board
Board upon which to flip pieces

-> [Position]
List of positions to flip

-> Board
Returns new board with pieces flipped

Flips all the pieces in the given list of Positions

flipPieces

Arguments

:: [(Position, Col)]
The list of pieces on the board

-> [Position]
The list of positions with pieces to be flipped

-> [(Position, Col)]
Returns a list of pieces with pieces flipped

Flips pieces on the board based on the given list of positions

flipPiece

Arguments

:: [(Position, Col)]
List of pieces on the board

-> Position
Piece to flip

-> [(Position, Col)]
Returns a new board with the piece flipped

Flips a single piece on the board

checkFlips

Arguments

:: [Position]
The list of positions that have pieces to be flipped

-> Board
The Board to check if positions contain pieces and the colours of pieces

-> Position
The Position where the move has been made

-> Position
The position offset used to check in a specific direction (eg, (0,-1) for North, (1,-1) for NorthEast)

-> Col
The Col which we are flipping the pieces to

-> [Position]
The return type. The fuction either returns its first argument or an empty list

Checks the board for any pieces that would be flipped

containsPiece

Arguments

:: Board
Board to check

-> Position
Positon to check for containing a piece

-> Bool
Returns true if position contains a piece and false otherwise

Checks that the given position of the board contains a piece

getPieceColor

Arguments

:: Board
The board to check

-> Position
The position on the board to check

-> Col
Returns the colour of the piece at this spot

Gets the colour of the piece at the given position. Assume that a piece exists in that position

checkScore

Arguments

:: Board
The board on which to check the current score

-> (Int, Int)
Returns tuple with black score and white core

Check the current score Returns a pair of the number of black pieces, and the number of white pieces

gameOver

Arguments

:: Board
Board to check for game being over

-> Bool
Return true of game is over and false otherwise

Return true if the game is complete (that is, either the board is full or there have been two consecutive passes)

evaluate

Arguments

:: Board
The board to evaluate

-> Col
The colour to evaluate the board for

-> Int
Returns number of pieces on boards for chosen colour

An evaluation function for a minimax search. Given a board and a colour return an integer indicating how good the board is for that colour.

undoTurn

Arguments

:: World
The world to be reverted to the previuos turn

-> World
returns the world in its previos turn

Reverts world back to most recent move - only human player turns are recorded and reverted to