Datatype

Documentation

type Position = (Int, Int)

data Col

A Col is either Black or White and represents the two opposing players

Constructors

Black

White

data Board

A Board is a record containing the board size (a board is a square grid, n * n), the number of consecutive passes, and a list of pairs of position and the colour at that position.

Constructors

Board

Fields

size :: Int

passes :: Int

pieces :: [(Position, Col)]

data PlayerType

PlayerType represents whether the player is Human or AI (more types can be added in future for different AI types)

Constructors

Human

AI

Random

Network

data World

Constructors

World

Fields

board :: Board

turn :: Col

stateList :: [(Board, Col, Int, Int)]

bType :: PlayerType

wType :: PlayerType

bTimer :: Int

wTimer :: Int

isPaused :: Bool

showValid :: Bool

chooseStart :: Bool

gameIsOver :: Bool

serverSide :: Bool

sock :: Maybe Socket