Draw

Documentation

data Bitmaps

Constructors

Bitmaps
Bitmaps are the data structure that stores all the bitmap files at the start of the game so they can be accessed by other draw functions

Fields

bp :: Picture

wp :: Picture

hp :: Picture

tp :: Picture

gob :: Picture

gow :: Picture

gobt :: Picture

gowt :: Picture

draw :: Picture

pause :: Picture

initPictures :: IO Bitmaps

Load images at the start of the game

width :: Float

height :: Double

gridPos :: Float

defaultSquareWidth :: Float

scaleFactor

Arguments

:: Int
Size of the board

-> Float

Calculates the factor the bitmaps have to be scaled depending on the board size

rectSize

Arguments

:: Int
The sz of the board/number of squares in a row

-> Float
returns the width of a single box

Calculates the width of a single square based on board sz

pieceSize

Arguments

:: Int
The sz of the board/number of squares in a row

-> Float
returns the radius of a piece

Calculates the radius of a piece based on the board sz

getCol :: Col -> Color

Translate between our custom Col and the Color needed for drawing

squarePosX

Arguments

:: Int
Size of the grid

-> Float
The position to be converted

-> Float
Returns the converted position

Converts an x position to an x position on the GUI for squares

squarePosY

Arguments

:: Int
Size of the grid

-> Float
The position to be converted

-> Float
Returns the converted position

Converts an y position to an y position on the GUI for squares

hintFunc :: Bool -> Col -> Board -> [Position]

drawWorldIO

Arguments

:: Bitmaps
Data structure with loaded bitmaps stored

-> World
The world to be drawn

-> IO Picture

Main draw world function passed into playIO

timeLeft

Arguments

:: PlayerType
Only draw if the playertype is human

-> Int
The time left

-> (Float, Float)
The (x,y) offset where the time should be drawn on the screen

-> IO Picture

Draw the time left for each player

drawBoardBMP

Arguments

:: Picture
Picture of the board tile

-> Int
The size of the board

-> IO Picture

Draws the empty board

getWinner

Arguments

:: Bitmaps
The saved bitmaps

-> Board
The board to evaluate to find winner

-> Int
The current time left for black

-> Int
The current time left for white

-> Picture

Gets a string stating the outcome of a game based on a board

drawText

Arguments

:: String
String to draw

-> Float
x coordinate to draw at

-> Float
y coordinate to draw at

-> IO Picture

Draws text

drawHints

Arguments

:: Picture
Picture for the individual hint pieces

-> Int
The size of the board on which hints are being drawn

-> [Position]
The list of positions to put hint pc

-> IO Picture

Draw hint pc on the board

drawEmptyPiece

Arguments

:: Picture
Picture for the individual spaces

-> Int
The size of the board

-> Position
The position draw the empty space on the board

-> IO Picture

Draw an empty space onto the board

drawPiecesBMP

Arguments

:: Picture
Picture for the black piece

-> Picture
Picture for the white piece

-> Int
The size of the board

-> [(Position, Col)]
List of positions and colours for drawing pc

-> IO Picture

Draws pc onto the boards

drawPieceBMP

Arguments

:: Picture
Picture for the black piece

-> Picture
Picture for the white piece

-> Int
The size of the board

-> Float
x position to draw piece

-> Float
y position to draw piece

-> Col
colour of piece that is to be drawn

-> IO Picture

Draw a single piece onto the board