Input

Documentation

getMover

Arguments

:: Bool
Boolean representing whether starting pieces are being placed

-> Board -> Position -> Col -> Maybe Board
movement function

Selects movement function to use based on whether placing start pieces or not

handleInputIO

Arguments

:: Event
Event to react to

-> World
World to update

-> IO World
Returns updated world

Update the world given an input, the event is traced in console window

snapX

Arguments

:: Int
The width of the board in squares

-> Float
the positon to convert to grid coordinate

-> Int
Returns the grid coordinate

snapY

Arguments

:: Int
The width of the board in squares

-> Float
The position to convert to grid coordinates

-> Int
Returns the grid coordinate